Move labels into TextField component.
