+++ client/src/components/TextField/TextField.module.css
@@ -1,4 +1,4 @@
-.text-input {
+.input {
   position: relative;
   box-sizing: border-box;
   width: 100%;
@@ -23,16 +23,26 @@
     font-weight: 300;
     color: var(--light-gray);
   }
+}
 
+.label {
+  display: block;
+  margin-bottom: 4px;
+}
+
+.field {
   &_theme_default {
-    padding: 14px 16px;
-    margin-bottom: 1px;
-    border: 1px solid var(--dark-gray-100);
+    & .input {
+      padding: 14px 16px;
+      border: 1px solid var(--dark-gray-100);
+    }
   }
 
   &_theme_flat {
-    padding: 15px 15px 15px 60px;
-    border: none;
-    box-shadow: inset 0 -2px 1px rgb(0 0 0 / 0.03);
+    & .input {
+      padding: 15px 15px 15px 60px;
+      border: none;
+      box-shadow: inset 0 -2px 1px rgb(0 0 0 / 0.03);
+    }
   }
 }
